rename to linkedpage, split out meta page interface
