+++ pkg/linkedpage/multi_test.go
@@ -1,4 +1,4 @@
-package metapage
+package linkedpage
 
 import (
 	"reflect"
